Optimized creating the new grid for the next generation
Rather than updating the state (alive or dead) of each cell in the grid by checking each of the cells neighbours, the new algorithm traverses the grid once. If it comes across a cell that is alive, it increments the value at the same coordinate in another matrix that stores the number of alive cells around each cell. Afterwards, the state of each cell in the grid is updated by assessing the number of alive cells around that cell, which is stored in the counts matrix.